Rework example tests
They now include the possibility of failure! Previously we just called
listchannel. Now we call sendmsg with both a positive and a negative
case.

In order to get this working with POST params, I had to make VCR
sensitive to the body of requests, in addition to the method and URI.

Also fixed a couple of grammar and formatting things.