Finish updating the Models lecture to match the live code
I know that @cheezitman has been working on the models lecture also.
I've tried to at least capture the spirit of his work, if not the text.
The bits on model.initialize() and get()/set()/unset() are all based off
what he wrote.

The first half of the lecture is about replacing our raw javascript
objects with real models, and it involves a pretty substantial amount of
refactoring. I've included a link to a diff on github that may be useful
in making those changes precise.

Again, moments to stop for discussion and check-in points are included.