Finish the Backbone Collections lecture
I've moved Chris's details on Collection functions to another file. This
stuff is valuable and should definitely be addressed, but I don't think
we'll have time tomorrow. It also doesn't fit well with the live code,
and would feel tacked-on if we left it there - it really deserves its
own space.