Suggestion - Reorder Intro to Ruby a bit
Small issue since it's a review of Jumpstart, but we've got if statements in the [Iteration & Blocks](https://github.com/Ada-Developers-Academy/textbook-curriculum/blob/master/01-ruby-fundamentals/iteration-and-blocks.md) lecture before we cover comparison & flow control.  Same thing for the 

Don't know why I didn't notice it before, probably because I was looking at the the lectures in isolation.