Rework example code in the class methods lecture
Used to be a hodge-podge of different examples, including the Penguin,
which isn't a realistic example of a program at all.

Now it's all focused on a Song class that represents a song in a music
player program. This should be a little more concrete, and the common
thread should make it easier to follow.