MC activity: Improve baseline setup instructions
We no longer instruct students to delete their scrap file after copying
its contents into the recipe file. This was unnecessary and at best it
added the potential for more confusion or un-intended merge conflicts.

Also slightly expanded on the instructions for distributing recipe
scraps and adding their contents to the recipe file.