Backbone live code: Genericize setup instructions
This is a difficult thing, and I'm not sure we have an ideal solution
yet. The problem is we should not link students to an AdaGold repo, but
we also shouldn't have our textbook curriculum link to a cohort-specific
repo.

For now I've changed the instructions to tell them to clone the repo
without giving an explicit link.