Remove the filtering-and-sorting resource
We talk about sorting already, and filtering isn't anything
backbone-specific, it's just some complex UI logic they could build. I
want to leave that as a challenge to them, especially because we're
supposed to be getting more hands-off at this point.

Also it doesn't really fit in the time alotted.