Update hash-tables.md
In this sentence, should it be O(n)? *It would be better to reduce the lookup time to a constant and the entire transaction to O(m).* I think it's unclear what the number of lookups means - that sounds like the number of times a person needs to look something up - and if each product  is an object with a unique id, the number of these objects is the defining constraint.