Update heaps.md
In the delete section, there was no step for actually deleting the node. I put that at step 2 but I'm not sure if that's the correct order. Graphics also need to be updated. Currently the last step says that the final heap should be slightly smaller, presumably because one element should be removed, but the graphic does not reflect that.